Updated Rendezvous3DOF
Added the capability to orient the corridor in any direction and also to rotate it over time (with a constant angular rate). Trained a new model with the corridor pointed in the +y direction (named it rdv_model_y). Note: this commit is being done from the command line because Github desktop is not responding.