Added option to save plots
Plots generated with plot_trajectory will now be saved in /plots if the --save flag is called. Animations are saved as html, and 2d plots are saved as png. Also, the plots will not be displayed unless the --show flag is called.
Also modified the --type argument to create all plots (animation & 2d) by default.
Removed plot_3d because it is not being used (code is still available in scratch file 'plotting').
Fixed some docstrings.