Added corridor randomizer to Rendezvous3DOF
The orientation of the entry corridor can now be randomized.
Added optional parameters to set the initial conditions for the environment.
Updated the render() function to show the corridor.
Made minor fixes to the Attitude environment.
Created eval_ scripts to evaluate the performance of the controllers on each environment.
Created utils files to store general functions.