Created animation with VPython
rendezvous_plot_vpython creates an animation of the chaser and the target. I switched from plotly to vpython because vpython can make the animation much quicker. However, the rotations in vpython are somewhat complicated, and I still need to figure out how to reset the attitude of the objects. The plotly animation can still be found in rendezvous_plot_plotly.
Utils:
Added a load_data() function to general utils (did not remove the one in animations utils).
Modified quaternion utils to directly normalize quaternions instead of checking their magnitude.
Added a vpython_utils module to store all the functions that are used for vpython animations.