Improved vpython animation
The animation made by the script rendezvous_plot_vpython now shows the rotation of both the chaser and the target.
Also, the animation can now be reset. The animation is automatically paused when it reaches the last timestep, and the play/pause button gets re-labeled to 'Reset'. Pressing the button moves the target and the chaser back to their original position and orientation.