Minor updates to RendezvousEnv
Changed the action from torque into an instantaneous change in rotational velocity.
Updated the collision check to account for the entry corridor.
Added coefficients to each term in the reward function. Also multiplied the velocity and rotation terms with a sigmoid function (because they should only be taken into account once the chaser is near the desired position).