Updated custom callback
`CustomWandbCallback` now also includes the environment as its argument. This will be necessary to evaluate the proper environment when tuning the reward function. I still need to add the tunable arguments to the environment itself.
Also separated the tuning script into three: one for algorithm, one for the policy, and one for the reward function.