Updated environment
Reverted back to the bubble reward (potential reward was feasible, but slow to learn - see `model_new`), and increased the time interval to 1 second. This should help to speed-up learning. The control inputs and rewards have been scaled by the time interval, which should allow us to switch back to the shorter time interval later.
Also added `total_delta_v` to the W&B logs (to track the fuel efficiency of the trajectories).