Updated callback
The evaluation function in `CustomWandbCallback` now records more characteristics of the trajectory, including collision percentage, time of first collision, and minimum position error. This was done because the previous `success` metric always returns zero (it is hard to achieve, need other metrics to show intermediate progress). The best result is no longer logged onto the W&B summary, instead it is recorded as an attribute of the Callback object, and it is only logged at the end of the training.