Update custom_callbacks.py
Changed the way in which the average of the minimum position errors and the collision times are recorded. The average of the minimum position errors will ignore any episodes where the chaser started inside the KOZ. And the average of the collision times will ignore any episodes where no collisions occurred. Previously this was not the case, and it was throwing-off the average.