Compute better initial attitude for target
Wrote a script `wait_for_target.py` that shows how to determine which initial attitude of the target makes it easier for the chaser to approach the entry corridor. The chaser could wait for this attitude and then begin its approach.
Also added some other ideas that could be implemented.