Implemented the max_reward model in NewEnv
There is now a maximum reward that can be achieved during a single episode. If the chaser achieves the terminal conditions for enough consecutive time steps, then the maximum reward is awarded for that episode.
This requires the success to be computed by the environment, so it is no longer computed by the callback (it is still recorded though)