Updated the schedule function
schedule_fn() is used to change the learning_rate as the training progresses. I added an optional parameter 'const' which determines if the learning rate is kept constant or not. If 'const' is True (by default), the learning rate does not change. Otherwise, the learning rate decreases linearly to zero as the training progresses.